various: Avoid using 'char' as an index
Depending on how the ctype macros are implemented and the compiler
warnings enabled, passing a char type the these macros may result in
the compiler complaining about it, given that char does not define
whether it's signed or unsigned.

+++ src/lib/common/sol-platform.c
@@ -158,7 +158,7 @@ sol_board_name_is_valid(const char *name)
         return false;
 
     for (i = 0; name[i] != '\0'; i++) {
-        if (isalnum(name[i]) || name[i] == '_' || name[i] == '-')
+        if (isalnum((uint8_t)name[i]) || name[i] == '_' || name[i] == '-')
             continue;
 
         return false;
